Изменено название package'й: вместо "com.example.myvoenmeh" - "com.my.voenmeh"
