+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
+
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
@@ -11,18 +12,20 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        tools:targetApi="31"
-        android:theme="@style/Theme.MyVoenmeh">
+        android:theme="@style/Theme.MyVoenmeh"
+        tools:targetApi="31">
+
         <activity
             android:name=".Activities.MainActivity"
             android:exported="false" />
         <activity
             android:name=".Activities.LoginActivity"
             android:exported="true"
-            android:theme="@style/loginPage"
-            android:label="@string/app_name">
+            android:label="@string/app_name"
+            android:theme="@style/loginPage">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
