--- app/src/main/java/com/example/myvoenmeh/Request/PostRequest.java
@@ -1,36 +1,0 @@
-package com.example.myvoenmeh.Request;
-
-import java.io.BufferedReader;
-import java.io.DataOutputStream;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.MalformedURLException;
-import java.net.URL;
-import javax.net.ssl.HttpsURLConnection;
-
-public class PostRequest {
-    public static void makeRequest() throws IOException {
-        String body = "{\"name\": \"Apple iPad Air\", \"data\": { \"Generation\": \"4th\", \"Price\": \"519.99\", \"Capacity\": \"256 GB\" }}";
-
-        URL url = new URL("https://api.restful-api.dev/objects");
-        HttpsURLConnection conn = (HttpsURLConnection) url.openConnection();
-        conn.setRequestMethod("POST");
-        conn.setDoOutput(true);
-        conn.setRequestProperty("Content-Type", "application/json");
-        conn.setRequestProperty("User-Agent", "Mozilla/5.0");
-
-        try (DataOutputStream dos = new DataOutputStream(conn.getOutputStream())) {
-            dos.writeBytes(body);
-        }
-
-        try (BufferedReader bf = new BufferedReader(new InputStreamReader(conn.getInputStream()))) {
-            String line;
-            while ((line = bf.readLine()) != null) {
-                System.out.println(line);
-            }
-        }
-    }
-
-    public PostRequest() throws IOException {
-    }
-}
