--- app/src/main/java/com/example/myvoenmeh/ui/mail/MailViewModel.java
@@ -1,6 +1,0 @@
-package com.example.myvoenmeh.ui.mail;
-
-import androidx.lifecycle.ViewModel;
-
-public class MailViewModel extends ViewModel {
-}
