--- app/src/main/java/com/example/myvoenmeh/ui/news/NewsFragment.java
@@ -1,34 +1,0 @@
-package com.example.myvoenmeh.ui.news;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
-
-import com.example.myvoenmeh.databinding.FragmentNewsBinding;
-
-public class NewsFragment extends Fragment {
-
-    private FragmentNewsBinding binding;
-
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             ViewGroup container, Bundle savedInstanceState) {
-        NewsViewModel newsViewModel =
-                new ViewModelProvider(this).get(NewsViewModel.class);
-
-        binding = FragmentNewsBinding.inflate(inflater, container, false);
-        View root = binding.getRoot();
-        return root;
-    }
-
-    @Override
-    public void onDestroyView() {
-        super.onDestroyView();
-        binding = null;
-    }
-}
