--- app/src/main/java/com/example/myvoenmeh/ui/news/NewsViewModel.java
@@ -1,6 +1,0 @@
-package com.example.myvoenmeh.ui.news;
-
-import androidx.lifecycle.ViewModel;
-
-public class NewsViewModel extends ViewModel {
-}
