--- app/src/main/java/com/example/myvoenmeh/ui/schedule/ScheduleViewModel.java
@@ -1,6 +1,0 @@
-package com.example.myvoenmeh.ui.schedule;
-
-import androidx.lifecycle.ViewModel;
-
-public class ScheduleViewModel extends ViewModel {
-}
