--- app/src/main/java/com/example/myvoenmeh/ui/service/ServiceFragment.java
@@ -1,34 +1,0 @@
-package com.example.myvoenmeh.ui.service;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
-
-import com.example.myvoenmeh.databinding.FragmentServiceBinding;
-
-public class ServiceFragment extends Fragment {
-
-    private FragmentServiceBinding binding;
-
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             ViewGroup container, Bundle savedInstanceState) {
-        ServiceViewModel serviceViewModel =
-                new ViewModelProvider(this).get(ServiceViewModel.class);
-
-        binding = FragmentServiceBinding.inflate(inflater, container, false);
-        View root = binding.getRoot();
-        return root;
-    }
-
-    @Override
-    public void onDestroyView() {
-        super.onDestroyView();
-        binding = null;
-    }
-}
