--- app/src/main/java/com/example/myvoenmeh/ui/service/ServiceViewModel.java
@@ -1,6 +1,0 @@
-package com.example.myvoenmeh.ui.service;
-
-import androidx.lifecycle.ViewModel;
-
-public class ServiceViewModel extends ViewModel {
-}
