--- app/src/main/java/com/example/myvoenmeh/ui/tracker/TrackerViewModel.java
@@ -1,6 +1,0 @@
-package com.example.myvoenmeh.ui.tracker;
-
-import androidx.lifecycle.ViewModel;
-
-public class TrackerViewModel extends ViewModel {
-}
