+++ app/src/main/java/com/my/voenmeh/ui/mail/MailViewModel.java
@@ -1,0 +1,6 @@
+package com.my.voenmeh.ui.mail;
+
+import androidx.lifecycle.ViewModel;
+
+public class MailViewModel extends ViewModel {
+}
