+++ app/src/main/java/com/my/voenmeh/ui/news/NewsViewModel.java
@@ -1,0 +1,6 @@
+package com.my.voenmeh.ui.news;
+
+import androidx.lifecycle.ViewModel;
+
+public class NewsViewModel extends ViewModel {
+}
