+++ app/src/main/java/com/my/voenmeh/ui/schedule/ScheduleFragment.java
@@ -1,0 +1,33 @@
+package com.my.voenmeh.ui.schedule;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.my.voenmeh.databinding.FragmentScheduleBinding;
+
+public class ScheduleFragment extends Fragment {
+
+    private FragmentScheduleBinding binding;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        ScheduleViewModel scheduleViewModel =
+                new ViewModelProvider(this).get(ScheduleViewModel.class);
+
+        binding = FragmentScheduleBinding.inflate(inflater, container, false);
+        View root = binding.getRoot();
+        return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+}
