+++ app/src/main/java/com/my/voenmeh/ui/schedule/ScheduleViewModel.java
@@ -1,0 +1,6 @@
+package com.my.voenmeh.ui.schedule;
+
+import androidx.lifecycle.ViewModel;
+
+public class ScheduleViewModel extends ViewModel {
+}
