+++ app/src/main/java/com/my/voenmeh/ui/service/ServiceFragment.java
@@ -1,0 +1,33 @@
+package com.my.voenmeh.ui.service;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.my.voenmeh.databinding.FragmentServiceBinding;
+
+public class ServiceFragment extends Fragment {
+
+    private FragmentServiceBinding binding;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        ServiceViewModel serviceViewModel =
+                new ViewModelProvider(this).get(ServiceViewModel.class);
+
+        binding = FragmentServiceBinding.inflate(inflater, container, false);
+        View root = binding.getRoot();
+        return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+}
