+++ app/src/main/java/com/my/voenmeh/ui/service/ServiceViewModel.java
@@ -1,0 +1,6 @@
+package com.my.voenmeh.ui.service;
+
+import androidx.lifecycle.ViewModel;
+
+public class ServiceViewModel extends ViewModel {
+}
