+++ app/src/main/java/com/my/voenmeh/ui/tracker/TrackerViewModel.java
@@ -1,0 +1,6 @@
+package com.my.voenmeh.ui.tracker;
+
+import androidx.lifecycle.ViewModel;
+
+public class TrackerViewModel extends ViewModel {
+}
