+++ app/src/main/res/values/themes.xml
@@ -12,4 +12,9 @@
         <item name="android:windowLightNavigationBar">true</item>
         <item name="android:navigationBarColor">@color/white</item>
     </style>
+    <!-- Base application theme. -->
+    <style name="Base.Theme.MyVoenmeh" parent="Theme.Material3.DayNight.NoActionBar">
+        <!-- Customize your light theme here. -->
+        <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
+    </style>
 </resources>
